Moved ArtistProfile & HostProfile to components. Is not workingwhen importing to the Account page.
