+++ client/src/pages/Account.js
@@ -1,5 +1,18 @@
 import React from "react";
+// import ArtistProfile from "../components/ArtistProfile";
+// import HostProfile from "../components/HostProfile";
+
 const Account = () => {
-  return <div className="">We are going to crush this!</div>;
+
+  return ( 
+
+  <div className="">
+    We are going to crush this
+    {/* <ArtistProfile />
+    <HostProfile /> */}
+  </div>
+
+  );
 };
+
 export default Account;
